services/music/pyproject.toml
[project]
name = "lyrics-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "crawl4ai>=0.7.3",
    "cssselect>=1.3.0",
    "dotenv>=0.9.9",
    "fastapi[standard]>=0.116.1",
    "orjson>=3.11.3",
    "pydantic-settings>=2.10.1",
    "python-dotenv>=1.1.1",
    "python-youtube>=0.9.8",
    "ruff>=0.12.9",
    "unidecode>=1.4.0",
    "yt-dlp>=2025.9.23",
]

[tool.ruff.lint]
ignore = ["F401"]

[tool.ruff.format]
quote-style = "preserve"        # "single" | "double" | "preserve"
indent-style = "space"          # "space" | "tab"
line-ending = "auto"            # "auto" | "lf" | "crlf" | "cr"
docstring-code-format = true    # format code blocks inside docstrings
preview = false
